Create JSON Submission
I have added an R script to the Code folder, called "Create_JSON_CU_Warning.R" and it; this script takes in the sample model output, and creates the additional fields necessary for the submission; it then writes the submission to a .json file.

I have added a sample CU Count .csv file to the Data folder; this is a placeholder for the output precition of our models.